third_party/blink/web_tests/fast/events/popup-blocked-from-different-frames-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE

PASS win is non-null.
PASS win is null
